# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Landing Template Swiss API",
        "description": "Centralized API contract for the Swiss landing template. Keep controllers annotated to auto-update this spec.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Primary API server"
        }
    ],
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Show the authenticated user",
                "operationId": "showAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "Authenticated user's profile.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/AuthenticatedUser"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    },
                    "error_code": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AuthenticatedUser": {
                "description": "Representation of the authenticated user returned by the profile endpoint.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sofia Admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "sofia@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-10T12:45:12+00:00",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T08:00:00+00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-05T13:37:00+00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Use the auto-login helper or paste a Sanctum token (Bearer <token>)."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}